+++ container_test.go
@@ -257,6 +257,7 @@ func TestContainer_Resolve(t *testing.T) {
 		err = c.Provide(func() (*http.Server, error) {
 			return &http.Server{}, fmt.Errorf("server build failed")
 		})
+		require.NoError(t, err)
 		err = c.Provide(func(server *http.Server) string {
 			return "string"
 		})
@@ -797,6 +798,7 @@ func TestContainer_Tags(t *testing.T) {
 		err = c.Provide(func() *http.Server { return first }, di.WithName("first"))
 		require.NoError(t, err)
 		err = c.Provide(func() *http.Server { return second }, di.WithName("second"))
+		require.NoError(t, err)
 		var extracted *http.Server
 		err = c.Resolve(&extracted)
 		require.Error(t, err)
@@ -1309,6 +1311,11 @@ func TestContainer_Inject(t *testing.T) {
 		require.True(t, extracted)
 		var result *InjectableType
 		require.NoError(t, c.Resolve(&result))
+
+		mux := http.NewServeMux()
+		p := InjectableParameter{Skipped: mux}
+		require.NoError(t, c.Resolve(&p))
+		require.Equal(t, InjectableParameter{Skipped: mux}, p)
 	})
 
 	t.Run("resolving not provided injectable cause error", func(t *testing.T) {
@@ -1426,4 +1433,81 @@ func TestContainer_Cleanup(t *testing.T) {
 		c.Cleanup()
 		require.Equal(t, []string{"server", "mux"}, cleanupCalls)
 	})
+}
+
+func TestContainer_AddParent(t *testing.T) {
+	t.Run("provide ancestor and resolve in child", func(t *testing.T) {
+		papaw, err := di.New()
+		require.NoError(t, err)
+		require.NotNil(t, papaw)
+		parent, err := di.New()
+		require.NoError(t, err)
+		require.NotNil(t, parent)
+		child, err := di.New()
+		require.NoError(t, err)
+		require.NotNil(t, child)
+
+		require.NoError(t, parent.AddParent(papaw))
+		require.NoError(t, child.AddParent(parent))
+
+		conn1 := &net.TCPConn{}
+		require.NoError(t, papaw.Provide(func() *net.TCPConn { return conn1 }))
+
+		var conn *net.TCPConn
+		require.NoError(t, child.Resolve(&conn))
+		require.Equal(t, fmt.Sprintf("%p", conn1), fmt.Sprintf("%p", conn))
+	})
+
+	t.Run("resolve multiple type instances across ancestors", func(t *testing.T) {
+		papaw, err := di.New()
+		require.NoError(t, err)
+		require.NotNil(t, papaw)
+		parent, err := di.New()
+		require.NoError(t, err)
+		require.NotNil(t, parent)
+		child, err := di.New()
+		require.NoError(t, err)
+		require.NotNil(t, child)
+
+		require.NoError(t, parent.AddParent(papaw))
+		require.NoError(t, child.AddParent(parent))
+
+		conn1 := &net.TCPConn{}
+		conn2 := &net.TCPConn{}
+		conn3 := &net.TCPConn{}
+		require.NoError(t, papaw.Provide(func() *net.TCPConn { return conn1 }))
+		require.NoError(t, parent.Provide(func() *net.TCPConn { return conn2 }))
+		require.NoError(t, child.Provide(func() *net.TCPConn { return conn3 }))
+
+		var conns []*net.TCPConn
+		require.NoError(t, child.Resolve(&conns))
+		require.Len(t, conns, 3)
+		require.Equal(t, fmt.Sprintf("%p", conn1), fmt.Sprintf("%p", conns[0]))
+		require.Equal(t, fmt.Sprintf("%p", conn2), fmt.Sprintf("%p", conns[1]))
+		require.Equal(t, fmt.Sprintf("%p", conn3), fmt.Sprintf("%p", conns[2]))
+	})
+
+	t.Run("add parent errors", func(t *testing.T) {
+		parent, err := di.New()
+		require.NoError(t, err)
+		require.NotNil(t, parent)
+		err = parent.AddParent(parent)
+		require.Contains(t, err.Error(), "self cycle detected")
+		child, err := di.New()
+		require.NoError(t, err)
+		require.NotNil(t, child)
+		err = child.AddParent(parent)
+		require.NoError(t, err)
+		err = parent.AddParent(child)
+		require.Contains(t, err.Error(), "cycle detected")
+		err = child.AddParent(parent)
+		require.Contains(t, err.Error(), "parent already chained")
+		papaw, err := di.New()
+		require.NoError(t, err)
+		require.NotNil(t, papaw)
+		err = parent.AddParent(papaw)
+		err = papaw.AddParent(child)
+		require.Contains(t, err.Error(), "cycle detected")
+	})
+
 }